fork browser-focused JS client to play nice with RN
This was required to avoid having to wrangle the native dependencies for the LD React Native client. Outside of checking which side of a feature flag a user is in, we don't need special native functionality so this will do in the interim until we can update RN to match the version they currently target (0.59.9).

Some of the problems I hit pulling their RN client into invest / trade, include:
* <React/RCTBridge.h> not found in their Pod bridge file (invest)
* requirement to use_modular_headers for React & some LaunchDarkly sub-pod (trade)

In general it seems their latest version supports XCode 12 which we need, but it doesn't support our current version of React which is not 100% pod friendly yet.